+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 - Revamped `cipher` package
 - Added box algorithm to `cipher` package
+- Added regex support to `blacklist`
 
 ## [v0.3.1]
 - Added `blacklist` package